Unit conversion fixes
Mixing floats, doubles and integers is not good for project.

+++ UtH-Engine-0.11/vs2013/TestProject/include/Turret.hpp
@@ -6,7 +6,7 @@
 class Turret : public uth::Component
 {
 public:
-	Turret(int t, int orb, int orbPos);
+	Turret(int t, float orb, float orbPos);
 	~Turret();
 
 	void init(int t);
